+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 next version
 ------------
 
+Changes
+=======
+
+* Added replacement of deterministic random variables (0 FIX) in modeling.cleanup_model
+
 Bugfixes
 ========